Test/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==2.0.7
databases==0.9.0
bcrypt==3.2.0
python-jose==3.3.0
passlib==1.7.4
